fix: switch to u8 syscall inner type

+++ eth-riscv-runtime/src/lib.rs
@@ -43,7 +43,7 @@ use eth_riscv_syscalls::Syscall;
 
 pub fn return_riscv(addr: u64, offset: u64) -> ! {
     unsafe {
-        asm!("ecall", in("a0") addr, in("a1") offset, in("t0") u64::from(Syscall::Return));
+        asm!("ecall", in("a0") addr, in("a1") offset, in("t0") u8::from(Syscall::Return));
     }
     unreachable!()
 }
@@ -51,26 +51,26 @@ pub fn return_riscv(addr: u64, offset: u64) -> ! {
 pub fn sload(key: u64) -> u64 {
     let value: u64;
     unsafe {
-        asm!("ecall", lateout("a0") value, in("a0") key, in("t0") u64::from(Syscall::SLoad));
+        asm!("ecall", lateout("a0") value, in("a0") key, in("t0") u8::from(Syscall::SLoad));
     }
     value
 }
 
 pub fn sstore(key: u64, value: u64) {
     unsafe {
-        asm!("ecall", in("a0") key, in("a1") value, in("t0") u64::from(Syscall::SStore));
+        asm!("ecall", in("a0") key, in("a1") value, in("t0") u8::from(Syscall::SStore));
     }
 }
 
 pub fn call(addr: u64, value: u64, in_mem: u64, in_size: u64, out_mem: u64, out_size: u64) {
     unsafe {
-        asm!("ecall", in("a0") addr, in("a1") value, in("a2") in_mem, in("a3") in_size, in("a4") out_mem, in("a5") out_size, in("t0") u64::from(Syscall::Call));
+        asm!("ecall", in("a0") addr, in("a1") value, in("a2") in_mem, in("a3") in_size, in("a4") out_mem, in("a5") out_size, in("t0") u8::from(Syscall::Call));
     }
 }
 
 pub fn revert() -> ! {
     unsafe {
-        asm!("ecall", in("t0") u64::from(Syscall::Revert));
+        asm!("ecall", in("t0") u8::from(Syscall::Revert));
     }
     unreachable!()
 }
@@ -90,7 +90,7 @@ pub fn keccak256(offset: u64, size: u64) -> B256 {
             lateout("a1") second,
             lateout("a2") third,
             lateout("a3") fourth,
-            in("t0") u64::from(Syscall::Keccak256)
+            in("t0") u8::from(Syscall::Keccak256)
         );
     }
 
@@ -109,7 +109,7 @@ pub fn msg_sender() -> Address {
     let second: u64;
     let third: u64;
     unsafe {
-        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, in("t0") u64::from(Syscall::Caller));
+        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, in("t0") u8::from(Syscall::Caller));
     }
     let mut bytes = [0u8; 20];
     bytes[0..8].copy_from_slice(&first.to_be_bytes());
@@ -124,7 +124,7 @@ pub fn msg_value() -> U256 {
     let third: u64;
     let fourth: u64;
     unsafe {
-        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u64::from(Syscall::CALLVALUE));
+        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u8::from(Syscall::CALLVALUE));
     }
     U256::from_limbs([first, second, third, fourth])
 }
@@ -133,7 +133,7 @@ pub fn msg_sig() -> [u8; 4] {
     let first: u64;
     let offset: u64 = 0;
     unsafe {
-        asm!("ecall", lateout("a0") first, in("a0") offset, in("t0") u64::from(Syscall::CALLDATALOAD));
+        asm!("ecall", lateout("a0") first, in("a0") offset, in("t0") u8::from(Syscall::CALLDATALOAD));
     }
     let mut bytes = [0u8; 4];
     bytes.copy_from_slice(&first.to_le_bytes()[0..4]);
@@ -144,7 +144,7 @@ pub fn msg_data() -> Vec<u8> {
     // Get the size of the call data
     let size: u64;
     unsafe {
-        asm!("ecall", lateout("a0") size, in("t0") u64::from(Syscall::CALLDATASIZE));
+        asm!("ecall", lateout("a0") size, in("t0") u8::from(Syscall::CALLDATASIZE));
     }
     // Load the call data
     let mut call_data = Vec::with_capacity(size as usize);
@@ -155,7 +155,7 @@ pub fn msg_data() -> Vec<u8> {
         let fourth: u64;
         // Load 32 bytes of call data
         unsafe {
-        asm!("ecall", in("a0") offset, lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u64::from(Syscall::CALLDATALOAD));
+        asm!("ecall", in("a0") offset, lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u8::from(Syscall::CALLDATALOAD));
         }
         let mut bytes = [0u8; 32];
         bytes[0..8].copy_from_slice(&first.to_le_bytes());
